CZB-504 use ownerOf function instead of zombieToOwner
