--- contracts/ZombieOwnership.sol
@@ -1,67 +1,0 @@
-// SPDX-License-Identifier: MIT
-pragma solidity ^0.8.16;
-
-import "./ZombieAttack.sol";
-import "./SafeMath.sol";
-
-/// @title A contract to make zombie tradable
-/// @author Shishir Sonekar
-/// @notice It follows all the necessary operations of ECR721 Standard
-abstract contract ZombieOwnership is ZombieAttack {
-    using SafeMath for uint256;
-
-    mapping(uint => address) zombieApprovals;
-
-    constructor(address _token) ZombieAttack(_token) {}
-
-    function balanceOf(address _owner)
-        public
-        view
-        override
-        returns (uint256 _balance)
-    {
-        return ownerZombieCount[_owner];
-    }
-
-    function ownerOf(uint256 _tokenId)
-        public
-        view
-        override
-        returns (address _owner)
-    {
-        return zombieToOwner[_tokenId];
-    }
-
-    function _transfer(
-        address _from,
-        address _to,
-        uint256 _tokenId
-    ) internal override {
-        ownerZombieCount[_to] = ownerZombieCount[_to].add(1);
-        ownerZombieCount[_from] = ownerZombieCount[_from].sub(1);
-        zombieToOwner[_tokenId] = _to;
-        emit Transfer(_from, _to, _tokenId);
-    }
-
-    function transfer(address _to, uint256 _tokenId)
-        public
-        onlyOwnerOf(_tokenId)
-    {
-        _transfer(msg.sender, _to, _tokenId);
-    }
-
-    function approve(address _to, uint256 _tokenId)
-        public
-        override
-        onlyOwnerOf(_tokenId)
-    {
-        zombieApprovals[_tokenId] = _to;
-        emit Approval(msg.sender, _to, _tokenId);
-    }
-
-    function takeOwnership(uint256 _tokenId) public {
-        require(zombieApprovals[_tokenId] == msg.sender);
-        address owner = ownerOf(_tokenId);
-        _transfer(owner, msg.sender, _tokenId);
-    }
-}
